Title: 이차원 배열의 오름차순 정리, Time: 15ms, Memory: 0MB, Status: Memory Limit Exceed - CodeTree

+++ 241106/이차원 배열의 오름차순 정리/ascending-order-of-two-dimensional-array.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 int n, k;
-vector<int> board;
+vector<long long> board;
 
 void Print()
 {
